Add the source from periods and time_for_keys
We will start with the sql from periods, remove the system/transaction
time support, and then refine to also pass the tests from time_for_keys.
The periods code also contains the changes from the PR that is supposed
to fix the inconsistencies in foreign key valid time checking, as well
as support for partitioned tables.

New tests are needed to verify the FK valid time algorithm,
and to verify the partitioned table support.