Fix trigger timing
All triggers should be deferrable, for allowing transactions
to have temporarily invalid states, that must be fixed by the end,
however, they should not be initially deferred, as that will
mask problems until the transaction completes.
Adjust the test accordingly, since it incorrectly dropped the
constraints before showing the table.