Adjust column names to start_after stop_on semantics
The sql_saga system supports different types, due to this
the name start is not appropriate, it is really start after, and stop is
stop on, so the previous stop column is identical to the next
start_after column, allowing for equality detection of same point.
With time this corresponds to valid_after and valid_to.